Shark and lights
I improved my shark mesh and texture(fixed errors, and removed some internal surfaces), added a resting position animation and added a non-Smooth version for if we want to go for the more polygon based style(currently both models are in). I also changed a setting in the rendering that fixed the problem of our models having weird shadows.